refactor: split `ls` and `cd` commands to different modules
