+++ fs/list-of-files.mjs
@@ -1,0 +1,41 @@
+import path from 'path';
+import { readdir, stat } from 'fs/promises'
+import { stdout } from 'process';
+import { EOL } from 'os';
+
+export const ls = async (currentPath) => {
+  const files = await readdir(currentPath)
+  const fileList = [];
+  const dirList = [];
+  const otherList = [];
+  for(let file of files) {
+    const pathToFile = path.join(currentPath, file)
+    try {
+      const stats = await stat(pathToFile)
+      if (file.length > 59) {
+        file = file.substring(0, 58) + '…';
+      }
+    
+      if(stats.isDirectory()) {
+        dirList.push(file);
+      } else {
+        fileList.push(file);
+      }
+    } catch (_) {
+      otherList.push(file);
+    }
+  }
+  console.log();
+  console.table( [
+    ...dirList
+      .sort((a, b) => a.toLowerCase().localeCompare(b.toLowerCase()))
+      .map((file) => ({Name: file, Type: 'directory'})),
+    ...fileList
+      .sort()
+      .map((file) => ({Name: file, Type: 'file'})),
+    ...otherList
+      .sort((a, b) => a.toLowerCase().localeCompare(b.toLowerCase()))
+      .map((file) => ({Name: file, Type: 'unknown'})),
+  ] );
+  stdout.write(EOL)
+}
